+++ tests/testthat/test-summarize_variables.R
@@ -127,6 +127,43 @@ test_that("s_summary works with length 0 factors that have levels", {
   expect_identical(result, expected)
 })
 
+test_that("s_summary works with factors and different denominator choices", {
+
+  x <- factor(c("Female", "Male", "Female", "Male", "Male", "Unknown", "Unknown", "Unknown", "Unknown"))
+
+  result <- s_summary(x, denom = "N_row", .N_row = 20)
+  expected <- list(
+    n = with_label(9L, "n"),
+    count_fraction = list(
+      Female = c(2, 2 / 20),
+      Male = c(3, 3 / 20),
+      Unknown = c(4, 4 / 20)
+    )
+  )
+  expect_identical(result, expected)
+
+  result <- s_summary(x, denom = "N_col", .N_col = 30)
+  expected <- list(
+    n = with_label(9L, "n"),
+    count_fraction = list(
+      Female = c(2, 2 / 30),
+      Male = c(3, 3 / 30),
+      Unknown = c(4, 4 / 30)
+    )
+  )
+  expect_identical(result, expected)
+})
+
+test_that("s_summary works with characters by converting to character", {
+
+  x <- c("Female", "Male", "Female", "Male", "Male", "Unknown", "Unknown", "Unknown", "Unknown")
+
+  result <- expect_warning(s_summary(x, denom = "N_row", .N_row = 20))
+  expected <- s_summary(factor(x), denom = "N_row", .N_row = 20)
+
+  expect_identical(result, expected)
+})
+
 test_that("`summarize_vars` works with healthy input, default `na.rm = TRUE`.", {
 
   dta_test <- data.frame(AVAL = c(1:4, NA, NA))
@@ -179,4 +216,69 @@ test_that("`summarize_vars` works with healthy factor input", {
   )
 
   expect_identical(result_matrix, expected_matrix)
+})
+
+test_that("`summarize_vars` works with factors and different denominators", {
+
+  start <- basic_table() %>%
+    split_cols_by("ARM") %>%
+    add_colcounts() %>%
+    split_rows_by("SEX", split_fun = drop_split_levels) %>%
+    add_rowcounts()
+
+  result_n <- start %>%
+    summarize_vars("RACE", denom = "n") %>%
+    build_table(DM)
+
+  result_ncol <- start %>%
+    summarize_vars("RACE", denom = "N_col") %>%
+    build_table(DM)
+
+  result <- start %>%
+    summarize_vars("RACE", denom = "N_row") %>%
+    build_table(DM)
+
+  expect_false(identical(result_n, result_ncol))
+  expect_false(identical(result_n, result))
+  expect_false(identical(result_ncol, result))
+
+  result_matrix <- to_string_matrix(result)
+  expected_matrix <- structure(
+    c("", "", "F (N=187)", "n", "ASIAN", "BLACK OR AFRICAN AMERICAN",
+      "WHITE", "AMERICAN INDIAN OR ALASKA NATIVE", "MULTIPLE", "NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER",
+      "OTHER", "UNKNOWN", "M (N=169)", "n", "ASIAN", "BLACK OR AFRICAN AMERICAN",
+      "WHITE", "AMERICAN INDIAN OR ALASKA NATIVE", "MULTIPLE", "NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER",
+      "OTHER", "UNKNOWN", "A: Drug X", "(N=121)", "", "70", "44 (23.5%)",
+      "18 (9.6%)", "8 (4.3%)", "0 (0%)", "0 (0%)", "0 (0%)", "0 (0%)",
+      "0 (0%)", "", "51", "35 (20.7%)", "10 (5.9%)", "6 (3.6%)", "0 (0%)",
+      "0 (0%)", "0 (0%)", "0 (0%)", "0 (0%)", "B: Placebo", "(N=106)",
+      "", "56", "37 (19.8%)", "12 (6.4%)", "7 (3.7%)", "0 (0%)", "0 (0%)",
+      "0 (0%)", "0 (0%)", "0 (0%)", "", "50", "31 (18.3%)", "12 (7.1%)",
+      "7 (4.1%)", "0 (0%)", "0 (0%)", "0 (0%)", "0 (0%)", "0 (0%)",
+      "C: Combination", "(N=129)", "", "61", "40 (21.4%)", "13 (7%)",
+      "8 (4.3%)", "0 (0%)", "0 (0%)", "0 (0%)", "0 (0%)", "0 (0%)",
+      "", "68", "44 (26%)", "14 (8.3%)", "10 (5.9%)", "0 (0%)", "0 (0%)",
+      "0 (0%)", "0 (0%)", "0 (0%)"),
+    .Dim = c(22L, 4L),
+    .Dimnames = list(NULL, c("", "A: Drug X", "B: Placebo", "C: Combination"))
+  )
+  expect_identical(result_matrix, expected_matrix)
+})
+
+test_that("`summarize_vars` works with character input and gives the same result as with factor", {
+
+  dta <- data.frame(
+    foo = c("a", "b", "a"),
+    stringsAsFactors = FALSE
+  )
+
+  l <- basic_table() %>%
+    summarize_vars(vars = "foo")
+  result <- expect_warning(build_table(l, dta))
+
+  dta_factor <- dta %>%
+    dplyr::mutate(foo = factor(foo))
+  expected <- build_table(l, dta_factor)
+
+  expect_identical(result, expected)
 })